moved reference from webdata to domain classes
