+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Domain/Event.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
@@ -7,6 +8,7 @@ namespace QRCodeAuth_Web.Models.Domain
 {
 	public class Event
 	{
+		[Key]
 		public string ID { get; set; }
 		public Account Owner { get; set; }
 		public string Name { get; set; }
