+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Domain/User.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
@@ -7,6 +8,7 @@ namespace QRCodeAuth_Web.Models
 {
 	public class User
 	{
+		[Key]
 		public string ID { get; set; }
 		public string SchoolID { get; set; }
 		public string LastName { get; set; }
